Add error_priv module; export error enums from new module; make error private
