+++ protocols/gossipsub/src/gossip_promises.rs
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::error::ValidationError;
+use crate::error_priv::ValidationError;
 use crate::peer_score::RejectReason;
 use crate::MessageId;
 use libp2p_core::PeerId;
